+++ src/components/InputText/InputText.tsx
@@ -1,5 +1,6 @@
 // @ts-ignore: React is used implicitly in JSX
-import React, { useState } from 'react'; // eslint-disable-line @typescript-eslint/no-unused-vars
+import React, { useState } from 'react';
+import { SemanticColorToken } from 'src/utils/new-colors';
 import { CharacterCounter } from '../CharacterCounter/CharacterCounter';
 import { GradeBand } from 'src/enum/gradeband';
 
@@ -7,6 +8,7 @@ export type InputTextProps = {
   value: string | number;
   onChange?: (value: string) => void;
   placeholderText?: string | undefined;
+  color?: SemanticColorToken;
   correct?: boolean;
   incorrect?: boolean;
   answerShown?: boolean;
@@ -22,6 +24,7 @@ export function InputText({
   value = '',
   onChange,
   placeholderText,
+  color,
   correct,
   incorrect,
   answerShown,
@@ -33,11 +36,15 @@ export function InputText({
 }: InputTextProps) {
   const [isSelected, setIsSelected] = useState(false);
 
+  const shouldBeDisabled = correct || incorrect || answerShown;
+
   const inputStates = [
     correct && 'connect__feedback-correct',
     incorrect && 'connect__feedback-incorrect',
     answerShown && 'connect__feedback-shown',
     isSelected && 'connect__selected',
+    (disabled || shouldBeDisabled) && 'connect__disabled',
+    color && `connect__color-${color}`,
     charLimit && 'connect__input-character-count',
   ]
     .filter(Boolean)
@@ -54,8 +61,6 @@ export function InputText({
     inputAriaLabel += ', answer shown';
   }
 
-  const shouldBeDisabled = correct || incorrect || answerShown || disabled;
-
   const handleTextChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     if (onChange) {
       onChange(event.currentTarget.value);
@@ -66,14 +71,14 @@ export function InputText({
     <label className={`connect__icon-wrapper ${inputStates}`}>
       <input
         type={isNumber}
-        className={`connect__input ${inputStates} ${disabled ? 'connect__disabled' : ''}`}
-        disabled={shouldBeDisabled}
+        className={`connect__input ${inputStates}`}
         value={value}
         placeholder={placeholderText ? placeholderText : ''}
         onChange={handleTextChange}
         onMouseDown={() => setIsSelected(true)}
         onBlur={() => setIsSelected(false)}
         aria-label={inputAriaLabel}
+        disabled={disabled || shouldBeDisabled}
         data-testid={dataTestId}
       />
       {onClear && (