+++ src/components/Table/Table.stories.tsx
@@ -11,7 +11,7 @@ const meta: Meta<typeof Table> = {
   component: Table,
   tags: ['autodocs'],
   parameters: {
-    layout: 'fullscreen',
+    layout: 'centered',
   },
 } as Meta;
 
@@ -34,6 +34,7 @@ export const Default = Template.bind({});
 Default.args = {
   scrolling: true,
   stickyHeader: false,
+  theme: '',
   headers: ['Header 1', 'Header 2', 'Header 3', 'Header 4'],
   cells: [
     ['Row 1 - Col 1', 'Row 1 - Col 2', 'Row 1 - Col 3', 'Row 1 - Col 4'],
@@ -70,6 +71,13 @@ ScopeRow.args = {
   ],
 };
 
+export const TableTheme = Template.bind({});
+TableTheme.args = {
+  ...Default.args,
+  caption: 'Table Theme',
+  theme: 'primary',
+};
+
 export const NoHeader = Template.bind({});
 NoHeader.args = {
   ...Default.args,
@@ -85,16 +93,20 @@ MobileListView.args = {
   scrolling: false,
 };
 
+MobileListView.parameters = {
+  layout: 'fullscreen',
+};
+
 export const FillInTheBlank = Template.bind({});
 FillInTheBlank.args = {
   ...Default.args,
   caption: 'Fill in the Blank',
   headers: ['Question', 'Answer'],
   cells: [
-    ['The capital of France is', <InputText />],
-    ['The capital of Germany is', <InputText />],
-    ['The capital of Italy is', <InputText />],
-    ['The capital of Japan is', <InputText />],
+    ['The capital of France is', <InputText value="" />],
+    ['The capital of Germany is', <InputText value="" />],
+    ['The capital of Italy is', <InputText value="" />],
+    ['The capital of Japan is', <InputText value="" />],
   ],
 };