+++ src/components/Table/Table.tsx
@@ -1,10 +1,14 @@
 import React from 'react';
+import { SemanticColorToken } from '../../utils/new-colors';
 import { GradeBand } from '../../enum/gradeband';
 
 export type TableProps = {
   headers?: string[];
+  subheaders?: string[];
   cells: string[][];
+  theme?: SemanticColorToken;
   caption?: string;
+  colgroup?: boolean;
   scrolling?: boolean;
   stickyHeader?: boolean;
   scopeCol?: boolean;
@@ -16,8 +20,11 @@ export type TableProps = {
 
 export const Table: React.FC<TableProps> = ({
   headers = [],
+  subheaders = [],
   cells = [],
   caption,
+  theme,
+  colgroup = true,
   scrolling = true,
   stickyHeader,
   scopeCol,
@@ -34,11 +41,24 @@ export const Table: React.FC<TableProps> = ({
     .filter(Boolean)
     .join(' ');
 
+  const themeColor = theme ? `var(--connect__color-${theme}-400)` : '';
+
+  const themeColorSticky = stickyHeader && theme ? `var(--connect__color-${theme}-100)` : '';
+
+  // eslint-disable-next-line no-console
+  console.log(`themeColor: ${themeColor}`);
+
   return (
     <div
       className="connect__table-wrapper"
       role="region"
       aria-labelledby={caption ? 'caption' : undefined}
+      style={
+        {
+          '--connect__table-theme-base': themeColor,
+          '--connect__table-theme-base-sticky': themeColorSticky,
+        } as React.CSSProperties
+      }
     >
       <table className={classNames} data-testid={dataTestId}>
         {caption && (
@@ -46,12 +66,24 @@ export const Table: React.FC<TableProps> = ({
             {caption}
           </caption>
         )}
+        {colgroup && (
+          <colgroup>
+            {headers.map((_, index) => (
+              <col key={index} />
+            ))}
+          </colgroup>
+        )}
         <thead>
           <tr>
             {headers.map((header, headerIndex) => (
               <th scope={scopeCol ? 'scope=col' : undefined} key={headerIndex[headerIndex]}>
                 {header}
               </th>
+            ))}
+            {subheaders.map((subheader, subheaderIndex) => (
+              <td className="subheader" key={subheaderIndex[subheaderIndex]}>
+                {subheader}
+              </td>
             ))}
           </tr>
         </thead>